+++ playground/src/ep/index.ts
@@ -1,3 +1,4 @@
 import './vue-formless-aug'
 
 export { FormView } from './form-view'
+export { FormCell } from 'vue-formless'